pyproject.toml
[project]
name = "langchain-memory-demo"
version = "0.1.0"
description = "LangGraph Memory Demo - Short-term and Long-term memory implementation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-huggingface>=0.1.0",
    "langchain-chroma>=0.1.0",
    "chromadb>=0.4.0",
    "langgraph>=0.2.50",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.4.0",
    "pytest>=8.0.0",
]
